# docs/openapi/CampusHelpATW.openapi.yaml
openapi: 3.0.3
info:
  title: CampusHelpATW API
  version: 0.1.0
  description: |
    CampusHelpATW 微信小程序后端接口（FastAPI 实现基线）。
    实时聊天 WebSocket 端点：
    - `/api/v1/ws/tasks/{task_id}?token={jwt}`
    全局未读通知 WebSocket 端点：
    - `/api/v1/ws/me/notifications?token={jwt}`
    全局通道事件：
    - `unread_snapshot` / `chat_unread`（聊天未读）
    - `task_event`（任务关键节点）
servers:
  - url: http://127.0.0.1:3000
    description: Local
  - url: https://api.example.com
    description: Production (placeholder)
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/me/profile:
    patch:
      tags: [me]
      summary: 更新个人资料
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdateRequest'
      responses:
        '200':
          description: 更新结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /api/v1/tasks/{task_id}/chat:
    get:
      tags: [chats]
      summary: 获取任务聊天会话
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 会话信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/chats/{chat_id}/messages:
    get:
      tags: [chats]
      summary: 获取聊天消息
      parameters:
        - in: path
          name: chat_id
          required: true
          schema:
            type: integer
        - in: query
          name: cursor
          schema:
            type: integer
            default: 0
        - in: query
          name: page_size
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 消息列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags: [chats]
      summary: 发送聊天消息
      parameters:
        - in: path
          name: chat_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessageCreateRequest'
      responses:
        '200':
          description: 发送结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /api/v1/chats/{chat_id}/read:
    post:
      tags: [chats]
      summary: 标记聊天已读
      parameters:
        - in: path
          name: chat_id
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatReadMarkRequest'
      responses:
        '200':
          description: 标记已读结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/posts:
    get:
      tags: [posts]
      summary: 帖子列表
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 帖子列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags: [posts]
      summary: 发帖
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreateRequest'
      responses:
        '200':
          description: 发帖结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/posts/{post_id}:
    get:
      tags: [posts]
      summary: 帖子详情
      parameters:
        - in: path
          name: post_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 帖子详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/posts/{post_id}/comments:
    post:
      tags: [posts]
      summary: 评论帖子
      parameters:
        - in: path
          name: post_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCommentCreateRequest'
      responses:
        '200':
          description: 评论结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/posts/{post_id}/like:
    post:
      tags: [posts]
      summary: 点赞帖子
      parameters:
        - in: path
          name: post_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 点赞结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      tags: [posts]
      summary: 取消点赞
      parameters:
        - in: path
          name: post_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 取消点赞结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          nullable: true
      required: [code, message]
...
    UserProfileUpdateRequest:
      type: object
      properties:
        nickname:
          type: string
        avatar_url:
          type: string
        college_name:
          type: string
...
    ChatMessageCreateRequest:
      type: object
      required: [content]
      properties:
        message_type:
          type: string
          default: TEXT
        content:
          type: string
    ChatReadMarkRequest:
      type: object
      properties:
        last_read_message_id:
          type: integer
          minimum: 0
    PostCreateRequest:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
    PostCommentCreateRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
...